requirements.txt
requests>=2.28.0
pandas>=1.5.0
numpy>=1.21.0
torch>=1.12.0
torchvision>=0.13.0
Pillow>=9.0.0
datasets>=2.0.0
transformers>=4.20.0
git+https://github.com/openai/CLIP.git
scikit-learn>=1.1.0
tqdm>=4.64.0
matplotlib>=3.5.0
opencv-python>=4.5.0
lmdb>=1.3.0